backend/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
whitenoise
python-dotenv
dj-database-url
drf-spectacular
django-cors-headers
django-filter
Pillow
django-storages[boto3]
sentry-sdk
django-redis
